requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
python-dotenv>=1.0.0
requests>=2.31.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
google-generativeai>=0.3.0
numpy>=1.24.0
apscheduler>=3.10.0
